+++ CHANGELOG.md
@@ -37,7 +37,7 @@ This changelog aims to follow the guiding principles of
 - Filter context creation API specification.
 - Gaussian Blur API specification.
 - In the OpenCV HAL, cvtColor YUV2RGB_NV21 is multithreaded.
-- In the OpenCV HAL, minMaxIdx is multithreaded when index is not requested.
+- In the OpenCV HAL, minMaxIdx is multithreaded.
 - Improved performance of Compare Equal and Greater SC API.
 
 ### Removed